requirements.txt
flask
line-bot-sdk
ultralytics
pillow
torch
opencv-python-headless
gunicorn
torchvision